+++ lib/executable-config/lookup/default.nix
@@ -1,5 +1,7 @@
 { hostPlatform, ghc
-, mkDerivation, base, bytestring, filepath, stdenv, text, transformers
+, mkDerivation, base, base64-bytestring, bytestring, filepath, stdenv, text, transformers
+, transformers-base, ref-tf, primitive, monad-control, reflex, reflex-dom
+, jsaddle
 , android-activity ? null, jsaddle-wkwebview ? null, ghcjs-dom ? null
 }:
 let isAndroid = hostPlatform.libc == "bionic";
@@ -10,7 +12,7 @@ in mkDerivation {
   version = "0.1";
   src = ./.;
   libraryHaskellDepends = [
-    base bytestring filepath text transformers
+    base bytestring filepath text transformers transformers-base ref-tf primitive monad-control reflex reflex-dom jsaddle
   ] ++ (if isAndroid then [
     android-activity
   ] else if isIOS then [