+++ lib/executable-config/lookup/src-ios/Obelisk/ExecutableConfig/Lookup.hs
@@ -1,15 +1,17 @@
 {-# LANGUAGE LambdaCase #-}
 module Obelisk.ExecutableConfig.Lookup where
 
+import Data.ByteString (ByteString)
 import Data.Map (Map)
 import Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
 import Language.Javascript.JSaddle.WKWebView
+import System.FilePath.Posix
 
-import Obelisk.ExecutableConfig.Internal.ConfigDirectory
+import Obelisk.Configs.Internal.Directory
 
-getConfigs :: IO (Map Text Text)
+getConfigs :: IO (Map Text ByteString)
 getConfigs = mainBundleResourcePath >>= \case
   Nothing -> error "Could not get bundle resource path."
-  Just _ -> getConfigsFromDirectory "config"
+  Just p -> getConfigsFromDirectory $ T.unpack (T.decodeUtf8 p) </> "config"