+++ lib/executable-config/lookup/src/Obelisk/ExecutableConfig/Lookup.hs
@@ -1,9 +1,0 @@
-module Obelisk.ExecutableConfig.Lookup where
-
-import Data.Map (Map)
-import Data.Text (Text)
-
-import Obelisk.ExecutableConfig.Internal.ConfigDirectory
-
-getConfigs :: IO (Map Text Text)
-getConfigs = getConfigsFromDirectory "config"