+++ lib/run/src/Obelisk/Run.hs
@@ -34,6 +34,7 @@ import Data.Streaming.Network (bindPortTCP)
 import Data.String (fromString)
 import Data.Text (Text)
 import qualified Data.Text as T
+import qualified Data.Text.Encoding as T
 import Language.Javascript.JSaddle.Run (syncPoint)
 import Language.Javascript.JSaddle.WebSockets
 import Network.HTTP.Client (Manager, defaultManagerSettings, newManager)
@@ -48,7 +49,6 @@ import Network.WebSockets (ConnectionOptions)
 import Network.WebSockets.Connection (defaultConnectionOptions)
 import qualified Obelisk.Asset.Serve.Snap as Snap
 import Obelisk.Backend
-import Obelisk.ExecutableConfig.Backend
 import Obelisk.Frontend
 import Obelisk.Route.Frontend
 import Reflex.Dom.Core
@@ -75,35 +75,34 @@ run port serveStaticAsset backend frontend = do
   case checkEncoder $ _backend_routeEncoder backend of
     Left e -> hPutStrLn stderr $ "backend error:\n" <> T.unpack e
     Right validFullEncoder -> do
-      configs <- getComponentConfigs
+      publicConfigs <- getPublicConfigs
       backendTid <- forkIO $ handle handleBackendErr $ withArgs ["--quiet", "--port", show port] $ do
-        runBackendConfigsT (_componentConfigs_backend configs) $
-          _backend_run backend $ \serveRoute -> do
-            runSnapWithCommandLineArgs $ do
-              getRouteWith validFullEncoder >>= \case
-                Identity r -> case r of
-                  InL backendRoute :=> Identity a ->
-                    runBackendConfigsT (_componentConfigs_backend configs) $
-                      serveRoute $ backendRoute :/ a
-                  InR obeliskRoute :=> Identity a ->
-                    serveDefaultObeliskApp (mkRouteToUrl validFullEncoder) serveStaticAsset frontend (_componentConfigs_frontend configs) $ obeliskRoute :/ a
+        _backend_run backend $ \serveRoute -> do
+          runSnapWithCommandLineArgs $ do
+            getRouteWith validFullEncoder >>= \case
+              Identity r -> case r of
+                InL backendRoute :=> Identity a -> serveRoute $ backendRoute :/ a
+                InR obeliskRoute :=> Identity a ->
+                  serveDefaultObeliskApp (mkRouteToUrl validFullEncoder) serveStaticAsset frontend publicConfigs $ obeliskRoute :/ a
       let conf = defRunConfig { _runConfig_redirectPort = port }
-      runWidget conf (_componentConfigs_frontend configs) frontend validFullEncoder `finally` killThread backendTid
+      runWidget conf publicConfigs frontend validFullEncoder `finally` killThread backendTid
 
 -- Convenience wrapper to handle path segments for 'Snap.serveAsset'
 runServeAsset :: FilePath -> [Text] -> Snap ()
 runServeAsset rootPath = Snap.serveAsset "" rootPath . T.unpack . T.intercalate "/"
 
-getConfigRoute :: Map Text Text -> Either Text URI
+getConfigRoute :: Map Text ByteString -> Either Text URI
 getConfigRoute configs = case Map.lookup "common/route" configs of
-  Just r -> case URI.mkURI $ T.strip r of
-    Just route -> Right route
-    Nothing -> Left $ "Couldn't parse route as URI; value read was: " <> T.pack (show r)
-  Nothing -> Left $ "Couldn't find config file common/route; it should contain the site's canonical root URI" <> T.pack (show $ Map.keys configs)
+    Just r -> 
+      let stripped = T.strip (T.decodeUtf8 r)
+      in case URI.mkURI stripped of
+          Just route -> Right route
+          Nothing -> Left $ "Couldn't parse route as URI; value read was: " <> T.pack (show stripped)
+    Nothing -> Left $ "Couldn't find config file common/route; it should contain the site's canonical root URI" <> T.pack (show $ Map.keys configs)
 
 runWidget
   :: RunConfig
-  -> Map Text Text
+  -> Map Text ByteString
   -> Frontend (R route)
   -> Encoder Identity Identity (R (Sum backendRoute (ObeliskRoute route))) PageName
   -> IO ()
@@ -125,7 +124,7 @@ runWidget conf configs frontend validFullEncoder = do
 
 obeliskApp
   :: forall route backendRoute
-  .  Map Text Text
+  .  Map Text ByteString
   -> ConnectionOptions
   -> Frontend (R route)
   -> Encoder Identity Identity (R (Sum backendRoute (ObeliskRoute route))) PageName
@@ -133,9 +132,12 @@ obeliskApp
   -> Application
   -> IO Application
 obeliskApp configs opts frontend validFullEncoder uri backend = do
-  let entryPoint = do
-        removeHTMLConfigs
-        runFrontendWithConfigs configs validFullEncoder frontend
+  let mode = FrontendMode
+        { _frontendMode_hydrate = True
+        , _frontendMode_adjustRoute = False
+        }
+      entryPoint = do
+        runFrontendWithConfigsAndCurrentRoute mode configs validFullEncoder frontend
         syncPoint
   jsaddlePath <- URI.mkPathPiece "jsaddle"
   let jsaddleUri = BSLC.fromStrict $ URI.renderBs $ uri & uriPath %~ (<>[jsaddlePath])
@@ -155,7 +157,7 @@ obeliskApp configs opts frontend validFullEncoder uri backend = do
       _ -> backend req sendResponse
 
 renderJsaddleFrontend
-  :: Map Text Text
+  :: Map Text ByteString
   -> Cookies
   -> (route -> Text)
   -> route