+++ ChangeLog.md
@@ -3,7 +3,10 @@
 This project's release branch is `master`. This log is written from the perspective of the release branch: when changes hit `master`, they are considered released, and the date should reflect that release.
 
 ## Unreleased
-
+* Configs become ByteStrings.
+* FrontendConfigsT has been changed into ConfigsT and configs are made available via getConfig/getConfigs
+  * The frontend will still only have access to configs that are placed in config/frontend and config/common, while the backend has access to the entire contents of the config directory via `Obelisk.ExecutableConfig.Lookup.getConfigs`.
+* The backend no longer runs in BackendConfigsT.
 * Add tabulation package. See Data.Tabulation for details.
 * Add encoders for `DMap`, `HasFields` (cf. Data.Tabulation), and JSON.